Notifications

+++ common/server-api/src/main/java/ru/kima/sonar/common/serverapi/model/HistoricCandle.kt
@@ -1,25 +1,24 @@
 package ru.kima.sonar.common.serverapi.model
 
-import java.math.BigDecimal
 import kotlin.time.Instant
 
 data class HistoricCandle(
     /**
      * Цена открытия за 1 инструмент. Чтобы получить стоимость лота, нужно умножить на лотность инструмента
      */
-    val open: BigDecimal,
+    val open: Double,
     /**
      * Максимальная цена за 1 инструмент
      */
-    val high: BigDecimal,
+    val high: Double,
     /**
      * Минимальная цена за 1 инструмент
      */
-    val low: BigDecimal,
+    val low: Double,
     /**
      * Цена закрытия за 1 инструмент
      */
-    val close: BigDecimal,
+    val close: Double,
     /**
      * Объем торгов в лотах.
      */
